Add files via upload
Separating layout.h into *.h and *.c
Added makefile
Updated README
Some changes to booksystem.c
