+++ Makefile
@@ -1,0 +1,8 @@
+CC = gcc
+CCFLAGS = -Wall
+
+main: 
+	$(CC) $(CCFLAGS) booksystem.c layout.c -o BookSystem
+
+clean:
+	rm BookSystem
